+++ src/main/java/io/nagurea/smsupsdk/webhooks/create/common/WebhookType.java
@@ -1,0 +1,9 @@
+package io.nagurea.smsupsdk.webhooks.create.common;
+
+public enum WebhookType {
+    MO,
+    DLR,
+    STOP,
+    CLICKER,
+    BALANCE_ALERT
+}